+++ CHANGELOG.md
@@ -8,13 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- login with enter key possible
-- added login with enter key test
+
+- add support for logging-in with ENTER key
 
 ### Changes
 
 - use npm over bun
-- update mex-common to 0.62.0
+- update mex-common to 0.62.1
 
 ### Deprecated